requirements.txt
Jinja2==3.1.2
Markdown==3.4.1
python-frontmatter==1.0.0
python-slugify==6.1.2
PyYAML==5.1